[ui-storagebrowser] moves the api and types to storagebrowsr folder
