+++ desktop/core/src/desktop/js/apps/storageBrowser/FileChooserModal/FileChooserModal.tsx
@@ -27,8 +27,8 @@ import { i18nReact } from '../../../utils/i18nReact';
 import useDebounce from '../../../utils/useDebounce';
 import useLoadData from '../../../utils/hooks/useLoadData/useLoadData';
 
-import { BrowserViewType, ListDirectory } from '../../../reactComponents/FileChooser/types';
-import { LIST_DIRECTORY_API_URL } from '../../../reactComponents/FileChooser/api';
+import { BrowserViewType, ListDirectory } from '../types';
+import { LIST_DIRECTORY_API_URL } from '../api';
 import PathBrowser from '../../../reactComponents/PathBrowser/PathBrowser';
 
 import './FileChooserModal.scss';
