+++ desktop/core/src/desktop/js/apps/storageBrowser/StorageDirectoryPage/StorageBrowserActions/DeletionModal/DeletionModal.test.tsx
@@ -2,11 +2,8 @@ import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import DeletionModal from './DeletionModal';
-import { StorageDirectoryTableData } from '../../../../../reactComponents/FileChooser/types';
-import {
-  BULK_DELETION_API_URL,
-  DELETION_API_URL
-} from '../../../../../reactComponents/FileChooser/api';
+import { StorageDirectoryTableData } from '../../../types';
+import { BULK_DELETION_API_URL, DELETION_API_URL } from '../../../api';
 
 const mockFiles: StorageDirectoryTableData[] = [
   {
