+++ desktop/core/src/desktop/js/apps/storageBrowser/StorageDirectoryPage/StorageBrowserActions/RenameModal/RenameModal.tsx
@@ -15,10 +15,10 @@
 // limitations under the License.
 
 import React from 'react';
-import { StorageDirectoryTableData } from '../../../../../reactComponents/FileChooser/types';
+import { StorageDirectoryTableData } from '../../../types';
 import { i18nReact } from '../../../../../utils/i18nReact';
 import useSaveData from '../../../../../utils/hooks/useSaveData/useSaveData';
-import { RENAME_API_URL } from '../../../../../reactComponents/FileChooser/api';
+import { RENAME_API_URL } from '../../../api';
 import InputModal from '../../../InputModal/InputModal';
 
 interface RenameModalProps {
